+++ docs/CHANGES.md
@@ -20,6 +20,7 @@ dev
  * Title or tick format strings can now be empty (useful if you do not want either of them)
  * Default title format is now `{title} {config_name}`
  * Uses temp file for MP3 encoding for safe abort even if that file is expected to overwrite existing one
+ * `normalize` tools is now properly invoked on Windows
  
 v1.1.0 (2018-10-04)
 -------------------